:recycle: refactor: update the unit test file for usePrograms with rename and refactor
